+++ packages/expo-calendar/android/src/main/java/expo/modules/calendar/next/domain/repositories/extendedproperty/ExtendedPropertyCursorExtensions.kt
@@ -1,0 +1,19 @@
+package expo.modules.calendar.next.domain.repositories.extendedproperty
+
+import android.database.Cursor
+import android.provider.CalendarContract
+import expo.modules.calendar.next.domain.model.extendedproperty.ExtendedPropertyEntity
+import expo.modules.calendar.next.domain.repositories.getOptionalLong
+import expo.modules.calendar.next.domain.repositories.getRequiredString
+import expo.modules.calendar.next.domain.wrappers.EventId
+import expo.modules.calendar.next.domain.wrappers.ExtendedPropertyId
+
+fun Cursor.toExtendedPropertyEntity(eventId: EventId) = ExtendedPropertyEntity(
+  id = ExtendedPropertyId(
+    getOptionalLong(CalendarContract.ExtendedProperties._ID)
+      ?: throw IllegalStateException("extended property ID must not be null")
+  ),
+  eventId = eventId,
+  name = getRequiredString(CalendarContract.ExtendedProperties.NAME),
+  value = getRequiredString(CalendarContract.ExtendedProperties.VALUE)
+)