+++ packages/expo-calendar/android/src/main/java/expo/modules/calendar/next/exceptions/ExtendedPropertiesExceptions.kt
@@ -1,0 +1,9 @@
+package expo.modules.calendar.next.exceptions
+
+import expo.modules.kotlin.exception.CodedException
+
+class ExtendedPropertyNameNotSyncSafeException(message: String, cause: Throwable? = null) :
+  CodedException(message, cause)
+
+class ExtendedPropertyAccountMissingException(message: String, cause: Throwable? = null) :
+  CodedException(message, cause)